+++ src/Core/Resources/Directories.cs
@@ -1,0 +1,13 @@
+﻿
+namespace AstralKeks.Workbench.Core.Resources
+{
+    internal class Directories
+    {
+        public const string Workbench = ".Workbench";
+        public const string Config = "Config";
+        public const string Core = "Core";
+        public const string Command = "Command";
+        public const string Cmdlet = "Cmdlet";
+        public const string Source = "src";
+    }
+}