+++ src/Core/Resources/Files.cs
@@ -1,0 +1,15 @@
+﻿
+namespace AstralKeks.Workbench.Core.Resources
+{
+    internal class Files
+    {
+        public const string Application = "WB-Application.json";
+        public const string Toolkit = "WB-Toolkit.json";
+        public const string Workspace = "WB-Workspace.json";
+        public const string WorkspaceLauncher = "Workspace.bat";
+        public const string Gitignore = ".gitignore";
+        public const string ChangeLog = "CHANGELOG.md";
+        public const string Licence = "LICENCE.md";
+        public const string Readme = "README.md";
+    }
+}